% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{aur_browse}
\alias{aur_browse}
\title{Browse AURIN data catalogue.}
\usage{
aur_browse()
}
\description{
A convenience function that opens up AURIN data catelogue in your default
web browser. Use the data catalogue to select spatial datasets available
on AURIN. Any datasets with 'AURIN Open API ID' field can be downloaded
into your current R session with \code{aur_get()}.
}
\examples{
\dontrun{

aur_browse()
}
}
