% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{aur_search}
\alias{aur_search}
\title{Search AURIN API}
\usage{
aur_search(query)
}
\arguments{
\item{query}{a character.}
}
\value{
a data.frame with two columns: aurin_open_api_id, brief_description.
}
\description{
Search AURIN API
}
