% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurinapi-package.R
\docType{package}
\name{aurin-package}
\alias{aurin}
\alias{aurin-package}
\title{aurin: Access AURIN Datasets from R}
\description{
Access AURIN datasets from R. To use this package you will need to request for an API key from AURIN (it's free!).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asiripanich/aurin}
  \item Report bugs at \url{https://github.com/asiripanich/aurin/issues}
}

}
\author{
\strong{Maintainer}: Amarin Siripanich \email{amarin.siri@gmail.com}

}
\keyword{internal}
