% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurinapi-package.R
\docType{package}
\name{aurinapi-package}
\alias{aurinapi}
\alias{aurinapi-package}
\title{aurinapi: A Package for Downloading AURIN Datasets Through their API.}
\description{
A Package for Downloading AURIN Datasets Through their API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asiripanich/aurinapi}
  \item Report bugs at \url{https://github.com/asiripanich/aurinapi/issues}
}

}
\keyword{internal}
