% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{aurinapi_browse}
\alias{aurinapi_browse}
\title{Browse AURIN data catalogue.}
\usage{
aurinapi_browse()
}
\description{
Use the data catalogue to select spatial datasets available on AURIN. Any datasets
with 'AURIN Open API ID' field can be downloaded into your current r session
with \code{aurinapi_get()}.
}
\examples{

\dontrun{

aurinapi_browse()

}
}
