% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{aurinapi_get}
\alias{aurinapi_get}
\title{Fetch AURIN dataset.}
\usage{
aurinapi_get(open_api_id, crs = "EPSG:4326", params = NULL)
}
\arguments{
\item{open_api_id}{Character. You can find the layer names for AURIN’s datasets
by browsing the \href{https://aurin.org.au/resources/aurin-apis/aurin-open-api-and-r/}{Data Catalogue}.
The layer name is found in the ‘AURIN Open API ID’ field in a dataset’s metadata.}

\item{crs}{default as "EPSG:4326". The Coordinate Reference System you wish to use}

\item{params}{a named list that will be used when constructing the WFS request.}
}
\value{
an \code{sf::sf} object.
}
\description{
Query data using \href{https://docs.geoserver.org/latest/en/user/services/wfs/reference.html}{WFS getFeature}.
}
\examples{

\dontrun{
# follow the example in https://aurin.org.au/resources/aurin-apis/aurin-open-api-and-r/
aurinapi_register(username = "your-username", password = "your-password")
aurinapi_get("aurin:datasource-au_govt_dss-UoM_AURIN_national_public_toilets_2017")
}
}
