% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{aurinapi_register}
\alias{aurinapi_register}
\title{Create an AURIN authentication file.}
\usage{
aurinapi_register(
  username,
  password,
  add_to_renviron = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{username}{character. Your username.}

\item{password}{character. Your password.}

\item{add_to_renviron}{logical, default as \code{FALSE}. If \code{TRUE} then the file directory
of the authentication file will be added to your \code{.Renviron} file.}
}
\description{
Create an AURIN authentication file.
}
\examples{
\dontrun{

aurinapi_register("username", "password")

}
}
