% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{aurinapi_wfs_client_wrapper}
\alias{aurinapi_wfs_client_wrapper}
\title{aurinapi_wfs_client_wrapper}
\description{
An R6 object for wrapping a WFS Client object. It will be created each
the \code{aurinapi} package is being loaded. See \link[ows4R:WFSClient]{ows4R::WFSClient}.
}
