% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{browse_aurin_catalogue}
\alias{browse_aurin_catalogue}
\title{Browse AURIN data catalogue.}
\usage{
browse_aurin_catalogue()
}
\description{
Use the data catalogue to select spatial datasets available on AURIN. Any datasets
with 'AURIN Open API ID' field can be downloaded into your current r session
with \code{fetch_aurin()}.
}
\examples{

\dontrun{

browse_aurin_catalogue()

}
}
