% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs-client.R
\name{create_aurinapi_wfs_client}
\alias{create_aurinapi_wfs_client}
\alias{get_aurinapi_wfs_client}
\alias{get_aurinapi_wfs_client_wrapper}
\title{Create a AURIN WFS client object.}
\usage{
create_aurinapi_wfs_client()

get_aurinapi_wfs_client()

get_aurinapi_wfs_client_wrapper()
}
\value{
a \code{ows4R::WFSClient} R6 object.
}
\description{
Create a AURIN WFS client object.
}
