% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_aurin.R
\name{fetch_aurin}
\alias{fetch_aurin}
\title{Fetch AURIN dataset.}
\usage{
fetch_aurin(open_api_id)
}
\arguments{
\item{open_api_id}{Character. You can find the layer names for AURIN’s datasets
by browsing the \href{https://aurin.org.au/resources/aurin-apis/aurin-open-api-and-r/}{Data Catalogue}.
The layer name is found in the ‘AURIN Open API ID’ field in a dataset’s metadata.}
}
\value{
an \code{sf::sf} object.
}
\description{
Fetch AURIN dataset.
}
\examples{

\dontrun{
# follow the example in https://aurin.org.au/resources/aurin-apis/aurin-open-api-and-r/
setup_authentication(username = "your-username", password = "your-password")
fetch("aurin:datasource-au_govt_dss-UoM_AURIN_national_public_toilets_2017")
}
}
