% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_authentication}
\alias{setup_authentication}
\title{Create an AURIN authentication file.}
\usage{
setup_authentication(
  username,
  password,
  add_to_renviron = FALSE,
  overwrite = FALSE,
  save_dir = Sys.getenv("HOME")
)
}
\arguments{
\item{username}{character. Your username.}

\item{password}{character. Your password.}

\item{add_to_renviron}{logical, default as \code{FALSE}. If \code{TRUE} then the file directory
of the authentication file will be added to your \code{.Renviron} file.
in the \code{.dir}.}

\item{overwrite}{Logical. Default as \code{FALSE}. If \code{TRUE} the authentication file will be overwritten.}

\item{save_dir}{a custom directory where the authentication file will be saved at.
If not given it will be saved in your home directory, as returns by Sys.getenv("HOME").
Regardless, of the directory the file will be saved as "aurin_wfs_connection.xml".}
}
\description{
Create an AURIN authentication file.
}
\examples{
\dontrun{

setup_authentication("username", "password")

}
}
