% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurinapi-package.R
\docType{package}
\name{aurin-package}
\alias{aurin}
\alias{aurin-package}
\title{aurin: Access Datasets from the 'AURIN' API}
\description{
'AURIN' <https://aurin.org.au/resources/aurin-apis/> is "Australia's single largest resource for accessing clean, integrated, spatially enabled and research-ready data on issues surrounding health and wellbeing, socio-economic metrics, transportation, and land-use.". This package provides functions to download and search datasets from the AURIN API (it's free to use!).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asiripanich/aurin}
  \item Report bugs at \url{https://github.com/asiripanich/aurin/issues}
}

}
\author{
\strong{Maintainer}: Amarin Siripanich \email{amarin.siri@gmail.com}

}
\keyword{internal}
