% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurinapi-package.R
\docType{package}
\name{aurinapi-package}
\alias{aurinapi}
\alias{aurinapi-package}
\title{aurinapi: Access AURIN datasets from R}
\description{
Access AURIN datasets from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asiripanich/aurinapi}
  \item Report bugs at \url{https://github.com/asiripanich/aurinapi/issues}
}

}
\author{
\strong{Maintainer}: Amarin Siripanich \email{amarin.siri@gmail.com}

}
\keyword{internal}
