% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{aurinapi_meta}
\alias{aurinapi_meta}
\title{Get metadata of all AURIN datasets}
\usage{
aurinapi_meta(force = FALSE)
}
\arguments{
\item{force}{Logical. Default is \code{FALSE}. Loading metadata may takes a while,
so if it has been loaded before it will not load again unluss \code{force} is set
to \code{TRUE}.}
}
\value{
a data.frame with two columns -- aurin_open_api_id and title.
}
\description{
Get metadata of all AURIN datasets
}
\examples{

\dontrun{

# this may takes a while to return.
aurinapi_meta()

}
}
